pyproject.toml
[project]
name = "server-cm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain[google-genai]>=1.2.7",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
]
